Implement using symlinks to prevent input file issues
Add a step that creates symbolic links of input fastq files with unique names in the resources subdirectory. This prevents issues when input multiple files have the same basename, even if their absolute file paths are different. Unique symlinks names are created by adding a 'run' prefix to the fastq basenames. A run prefix can be provided in the optional 'run' column in read_files.tsv. If not provided, the pipeline will create an arbitrary run prefix for each read pair.